requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
aiofiles==23.2.1
    # via chainlit
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via
    #   asyncer
    #   httpx
    #   langfuse
    #   openai
    #   starlette
    #   watchfiles
asyncer==0.0.7
    # via chainlit
backoff==2.2.1
    # via langfuse
bidict==0.23.1
    # via python-socketio
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
chainlit==1.2.0
    # via -r requirements.in
charset-normalizer==3.3.2
    # via requests
chevron==0.14.0
    # via literalai
click==8.1.7
    # via
    #   chainlit
    #   uvicorn
dataclasses-json==0.6.7
    # via chainlit
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-semantic-conventions
distro==1.9.0
    # via openai
exceptiongroup==1.2.2
    # via anyio
fastapi==0.112.4
    # via chainlit
filetype==1.2.0
    # via chainlit
google-search-results==2.4.2
    # via -r requirements.in
googleapis-common-protos==1.65.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.66.1
    # via opentelemetry-exporter-otlp-proto-grpc
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
    #   wsproto
httpcore==1.0.5
    # via httpx
httpx==0.27.2
    # via
    #   chainlit
    #   langfuse
    #   langsmith
    #   literalai
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   langfuse
    #   requests
importlib-metadata==8.4.0
    # via opentelemetry-api
jiter==0.5.0
    # via openai
langfuse==2.50.2
    # via -r requirements.in
langsmith==0.1.125
    # via -r requirements.in
lazify==0.4.0
    # via chainlit
literalai==0.0.607
    # via chainlit
marshmallow==3.22.0
    # via dataclasses-json
mypy-extensions==1.0.0
    # via typing-inspect
nest-asyncio==1.6.0
    # via chainlit
numpy==1.26.4
    # via chainlit
openai==1.47.0
    # via -r requirements.in
opentelemetry-api==1.27.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   uptrace
opentelemetry-exporter-otlp==1.27.0
    # via uptrace
opentelemetry-exporter-otlp-proto-common==1.27.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.27.0
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.27.0
    # via opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.48b0
    # via uptrace
opentelemetry-proto==1.27.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.27.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   uptrace
opentelemetry-semantic-conventions==0.48b0
    # via opentelemetry-sdk
orjson==3.10.7
    # via langsmith
packaging==23.2
    # via
    #   chainlit
    #   langfuse
    #   literalai
    #   marshmallow
protobuf==4.25.5
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pydantic==2.9.2
    # via
    #   chainlit
    #   fastapi
    #   langfuse
    #   langsmith
    #   literalai
    #   openai
pydantic-core==2.23.4
    # via pydantic
pyjwt==2.9.0
    # via chainlit
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   chainlit
python-engineio==4.9.1
    # via python-socketio
python-multipart==0.0.9
    # via chainlit
python-socketio==5.11.4
    # via chainlit
requests==2.32.3
    # via
    #   google-search-results
    #   langsmith
    #   opentelemetry-exporter-otlp-proto-http
    #   serpapi
serpapi==0.1.5
    # via -r requirements.in
simple-websocket==1.0.0
    # via python-engineio
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
starlette==0.37.2
    # via
    #   chainlit
    #   fastapi
syncer==2.0.3
    # via chainlit
tomli==2.0.1
    # via chainlit
tqdm==4.66.5
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   asyncer
    #   fastapi
    #   openai
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspect
    #   uvicorn
typing-inspect==0.9.0
    # via dataclasses-json
uptrace==1.26.0
    # via chainlit
urllib3==2.2.3
    # via requests
uvicorn==0.25.0
    # via chainlit
watchfiles==0.20.0
    # via chainlit
wrapt==1.16.0
    # via
    #   deprecated
    #   langfuse
    #   opentelemetry-instrumentation
wsproto==1.2.0
    # via simple-websocket
zipp==3.20.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
